+++ mod-source-record-manager-server/src/main/java/org/folio/services/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(groupPackage = "org.folio.services", name = "journal-service")
-package org.folio.services;
-
-import io.vertx.codegen.annotations.ModuleGen;